# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Frame Control System API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /users/register:
    post:
      summary: Register user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
        '409':
          description: Email taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /users/login:
    post:
      summary: Login and get JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /users/me:
    get:
      summary: Get current user profile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
    patch:
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /users:
    get:
      summary: List users (admin)
      parameters:
        - in: query
          name: email
          schema: { type: string }
        - in: query
          name: name
          schema: { type: string }
        - in: query
          name: role
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, enum: [email_asc,email_desc,name_asc,name_desc,created_asc,created_desc] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /orders:
    get:
      summary: List orders
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [created,in_progress,done,cancelled] }
        - in: query
          name: sort
          schema: { type: string, enum: [created_asc,created_desc] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
    post:
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /orders/{id}:
    get:
      summary: Get order by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
    delete:
      summary: Cancel order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
  /orders/{id}/status:
    patch:
      summary: Update order status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
        '400':
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeError'
  /events/outbox:
    get:
      summary: List outbox events (admin)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOk'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EnvelopeOk:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: {}
    EnvelopeError:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    RegisterRequest:
      type: object
      required: [email,password,name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        name: { type: string }
    LoginRequest:
      type: object
      required: [email,password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    UpdateMeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
    OrderItem:
      type: object
      required: [name,quantity,price]
      properties:
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
        price: { type: number, minimum: 0 }
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [in_progress, done]