# internal/httpserver/server.go
package httpserver

import (
	"encoding/json"
	"net/http"
	"database/sql"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"

	"frame_control_system/internal/config"
)

type envelope struct {
	Success bool        `json:"success"`
	Data    interface{} `json:"data,omitempty"`
	Error   *apiError   `json:"error,omitempty"`
}

type apiError struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

func NewRouter(cfg config.Config, db *sql.DB) http.Handler {
	r := chi.NewRouter()

	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(middleware.Recoverer)
	r.Use(middleware.Timeout(30 * time.Second))

	corsMw := cors.Handler(cors.Options{
		AllowedOrigins:   cfg.CORSOrigins,
		AllowedMethods:   []string{"GET", "POST", "PATCH", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-Request-ID"},
		ExposedHeaders:   []string{"X-Request-ID"},
		AllowCredentials: false,
		MaxAge:           300,
	})
	r.Use(corsMw)

	// Project middlewares
	r.Use(Logger())
	r.Use(RateLimit(cfg.RateLimitRPS, cfg.RateLimitBurst))

	r.Route("/api/v1", func(v1 chi.Router) {
		v1.Get("/healthz", func(w http.ResponseWriter, r *http.Request) {
			writeJSON(w, http.StatusOK, envelope{Success: true, Data: map[string]string{"status": "ok"}})
		})

		// Auth
		v1.Post("/users/register", RegisterHandler(db))
		v1.Post("/users/login", LoginHandler(db, cfg.JWTSecret))

		// Protected
		v1.Group(func(pr chi.Router) {
			pr.Use(AuthMiddleware(cfg.JWTSecret))

			// Me
			pr.Get("/users/me", GetMeHandler(db))
			pr.Patch("/users/me", UpdateMeHandler(db))

			// Admin
			pr.With(RequireRole("admin")).Get("/users", AdminListUsersHandler(db))
			pr.With(RequireRole("admin")).Get("/events/outbox", AdminListOutboxHandler(db))

			// Orders
			pr.Post("/orders", CreateOrderHandler(db))
			pr.Get("/orders", ListOrdersHandler(db))
			pr.Get("/orders/{id}", GetOrderHandler(db)) // prefer path param
			pr.Patch("/orders/{id}/status", UpdateOrderStatusHandler(db))
			pr.Delete("/orders/{id}", CancelOrderHandler(db))
		})
	})

	r.NotFound(func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, http.StatusNotFound, envelope{Success: false, Error: &apiError{Code: "not_found", Message: "resource not found"}})
	})
	r.MethodNotAllowed(func(w http.ResponseWriter, r *http.Request) {
		writeJSON(w, http.StatusMethodNotAllowed, envelope{Success: false, Error: &apiError{Code: "method_not_allowed", Message: "method not allowed"}})
	})

	return r
}

func writeJSON(w http.ResponseWriter, status int, v interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}